implement basic roboblock application
Implement a very basic robocaller blocker application. This application gives each caller a trusted percentage. If they are not trusted they are presented a random captcha to solve. Upon completion their trust is boosted. On a fail their trust is deboosted. Above 80% trusted a caller is no longer presented with captchas.

This application still lacks the ability to manage settings such as:
 -  boost/deboost percentage  default settings
 -  proper ui to manage the settings such as manual block/allow
 -  keybindings that allow boosting/deboosting a caller while on the call with him.
 -  ability to completely block a caller in the future if their trust is too low